refactor: IntegratedDailyScheduleCell 위치 변경
