+++ frontend/src/components/IntegratedDailySchedule/IntegratedDailyScheduleCell/IntegratedDailyScheduleCell.tsx
